Add files via upload
The goal of this project is to build a predictive model that can accurately identify recipes that are likely to generate high traffic. By leveraging historical data on recipe features and traffic patterns, we can train a machine-learning model to predict the likelihood of a recipe being popular and generating high traffic.